Inventory manager has been updated to utilize enums where possible -- duplicate functions and matches have been created to ensure the transition can be made, just need to transition at this point.  The scene transition commands have been fully implemented and can be used with set scene <scene_name>

+++ Psyche/Assets/Scripts/Player/ToolManagers/ThrusterManager.cs
@@ -45,29 +45,29 @@ public class ThrusterManager : ToolManager
         toolEnabled = false;
         _playerController = playerController;
         level = 0;
-        levelRequirements = new Dictionary<int, Dictionary<string, int>>()
+        levelRequirements = new Dictionary<int, Dictionary<string, ushort>>()
         {
-            {  1, new Dictionary<string, int>()
+            {  1, new Dictionary<string, ushort>()
                 {
                     { "element_copper", 1 }, { "element_iron", 0 }, { "element_nickel", 0 }, { "element_gold", 0 }, { "element_platinum", 0 }
                 }
             },
-            {  2, new Dictionary<string, int>()
+            {  2, new Dictionary<string, ushort>()
                 {
                     { "element_copper", 2 } , { "element_iron", 0 } , { "element_nickel", 0 } , { "element_gold", 0 }, { "element_platinum", 0 }
                 }
             },
-            {  3, new Dictionary<string, int>()
+            {  3, new Dictionary<string, ushort>()
                 {
                     { "element_copper", 3 } , { "element_iron", 0 } , { "element_nickel", 0 } , { "element_gold", 0 }, { "element_platinum", 0 }
                 }
             },
-            {  4, new Dictionary<string, int>()
+            {  4, new Dictionary<string, ushort>()
                 {
                     { "element_copper", 4 } , { "element_iron", 0 } , { "element_nickel", 0 } , { "element_gold", 0 }, { "element_platinum", 0 }
                 }
             },
-            {  5, new Dictionary<string, int>()
+            {  5, new Dictionary<string, ushort>()
                 {
                     { "element_copper", 5 } , { "element_iron", 0 } , { "element_nickel", 0 } , { "element_gold", 0 }, { "element_platinum", 0 }
                 }
@@ -100,6 +100,4 @@ public class ThrusterManager : ToolManager
     {
         _thrusterForce += 0.25f;
     }
-
-
 }
